完善供应商发货

+++ model/src/main/java/cn/qihangerp/model/entity/OOrderItem.java
@@ -131,8 +131,11 @@ public class OOrderItem implements Serializable {
      */
     private Integer orderStatus;
 
-    //发货方式 0 自己发货1联合发货2供应商发货
-    private Integer shipper;
+    //发货人
+    private Long shipper;
+    /**
+     * 发货方式 0 自己发货1联合发货2供应商发货
+     */
     private Integer shipType;//发货方式1电子面单发货2手动发货
 
     /**
